+++ dart-ffi/src/link.rs
@@ -6,12 +6,12 @@ use isar_core::error::Result;
 pub unsafe extern "C" fn isar_link(
     collection: &'static IsarCollection,
     txn: &mut CIsarTxn,
-    link_id: u32,
+    link_id: u64,
     id: i64,
     target_id: i64,
 ) -> i64 {
     isar_try_txn!(txn, move |txn| -> Result<()> {
-        collection.link(txn, link_id as usize, id, target_id)?;
+        collection.link(txn, link_id, id, target_id)?;
         Ok(())
     })
 }
@@ -20,12 +20,12 @@ pub unsafe extern "C" fn isar_link(
 pub unsafe extern "C" fn isar_link_unlink(
     collection: &'static IsarCollection,
     txn: &mut CIsarTxn,
-    link_id: u32,
+    link_id: u64,
     id: i64,
     target_id: i64,
 ) -> i64 {
     isar_try_txn!(txn, move |txn| -> Result<()> {
-        collection.unlink(txn, link_id as usize, id, target_id)?;
+        collection.unlink(txn, link_id, id, target_id)?;
         Ok(())
     })
 }
@@ -34,11 +34,11 @@ pub unsafe extern "C" fn isar_link_unlink(
 pub unsafe extern "C" fn isar_link_unlink_all(
     collection: &'static IsarCollection,
     txn: &mut CIsarTxn,
-    link_id: u32,
+    link_id: u64,
     id: i64,
 ) -> i64 {
     isar_try_txn!(txn, move |txn| -> Result<()> {
-        collection.unlink_all(txn, link_id as usize, id)?;
+        collection.unlink_all(txn, link_id, id)?;
         Ok(())
     })
 }
@@ -47,7 +47,7 @@ pub unsafe extern "C" fn isar_link_unlink_all(
 pub unsafe extern "C" fn isar_link_update_all(
     collection: &'static IsarCollection,
     txn: &mut CIsarTxn,
-    link_id: u32,
+    link_id: u64,
     id: i64,
     ids: *const i64,
     link_count: u32,
@@ -57,17 +57,17 @@ pub unsafe extern "C" fn isar_link_update_all(
     let ids = std::slice::from_raw_parts(ids, (link_count + unlink_count) as usize);
     isar_try_txn!(txn, move |txn| {
         if replace {
-            collection.unlink_all(txn, link_id as usize, id)?;
+            collection.unlink_all(txn, link_id, id)?;
         }
         for target_id in ids.iter().take(link_count as usize) {
-            collection.link(txn, link_id as usize, id, *target_id)?;
+            collection.link(txn, link_id, id, *target_id)?;
         }
         for target_id in ids
             .iter()
             .skip(link_count as usize)
             .take(unlink_count as usize)
         {
-            collection.unlink(txn, link_id as usize, id, *target_id)?;
+            collection.unlink(txn, link_id, id, *target_id)?;
         }
         Ok(())
     })
